evaluator: remove Eval function.
